Create about-me.css
CSS for my about me page.  The pencils took some time to adjust correctly.  I wanted to go with a more simplistic color scheme and design.  Originally I was going to try to edit my website from Web Dev I but it was a bit too messy and felt better to start fresh.